requirements.txt
cassandra-driver
time_uuid
pydgraph
fastapi[all]
pydantic
pymongo
requests